# Conflicts:
#	src/main/java/S13P31A306/loglens/domain/log/controller/LogApi.java
#	src/main/java/S13P31A306/loglens/domain/log/controller/impl/LogController.java
#	src/main/java/S13P31A306/loglens/domain/log/dto/request/LogSearchRequest.java
#	src/main/java/S13P31A306/loglens/domain/log/service/LogService.java
#	src/main/java/S13P31A306/loglens/domain/log/service/impl/LogServiceImpl.java
#	src/main/resources/application.yml

+++ src/test/java/S13P31A306/loglens/domain/jira/service/impl/JiraIssueServiceImplTest.java
@@ -78,7 +78,7 @@ class JiraIssueServiceImplTest {
             Integer userId = 1;
             Integer projectId = 1;
             String projectUuid = "550e8400-e29b-41d4-a716-446655440000";
-            Long logId = 100L;
+            Integer logId = 100;
             String summary = "Test Issue";
             String description = "Test Description";
 
@@ -198,7 +198,7 @@ class JiraIssueServiceImplTest {
             String projectUuid = "550e8400-e29b-41d4-a716-446655440000";
             JiraIssueCreateRequest request = new JiraIssueCreateRequest(
                     projectUuid,
-                    100L,
+                    100,
                     "Test Issue",
                     "Test Description",
                     "Bug",
@@ -241,7 +241,7 @@ class JiraIssueServiceImplTest {
             Integer userId = 1;
             Integer projectId = 1;
             String projectUuid = "550e8400-e29b-41d4-a716-446655440000";
-            Long logId = 100L;
+            Integer logId = 100;
             JiraIssueCreateRequest request = new JiraIssueCreateRequest(
                     projectUuid,
                     logId,
@@ -288,7 +288,7 @@ class JiraIssueServiceImplTest {
             Integer userId = 1;
             Integer projectId = 1;
             String projectUuid = "550e8400-e29b-41d4-a716-446655440000";
-            Long logId = 100L;
+            Integer logId = 100;
             JiraIssueCreateRequest request = new JiraIssueCreateRequest(
                     projectUuid,
                     logId,
@@ -338,7 +338,7 @@ class JiraIssueServiceImplTest {
             Integer userId = 1;
             Integer projectId = 1;
             String projectUuid = "550e8400-e29b-41d4-a716-446655440000";
-            Long logId = 100L;
+            Integer logId = 100;
             JiraIssueCreateRequest request = new JiraIssueCreateRequest(
                     projectUuid,
                     logId,
